+++ CHANGELOG.md
@@ -2,8 +2,11 @@
 
 # Unreleased
 
+## Added
+
 - Add `--longest-line` flag.
 - Add `--include-longest-line` flag.
+- Add `--sort-by` and `--sort-order` flags.
 
 # 0.4.2